dolfin: allow int32 cell data (with test)

+++ test/helpers.py
@@ -225,7 +225,7 @@ def add_point_data(mesh, dim, num_tags=2):
     return mesh2
 
 
-def add_cell_data(mesh, dim, num_tags=2):
+def add_cell_data(mesh, dim, num_tags=2, dtype=numpy.float):
     mesh2 = _clone(mesh)
     numpy.random.seed(0)
     cell_data = {}
@@ -233,12 +233,14 @@ def add_cell_data(mesh, dim, num_tags=2):
         num_cells = len(mesh['cells'][cell_type])
         if dim == 1:
             cell_data[cell_type] = {
-                string.ascii_lowercase[k]: numpy.random.rand(num_cells)
+                string.ascii_lowercase[k]:
+                numpy.random.rand(num_cells).astype(dtype)
                 for k in range(num_tags)
                 }
         else:
             cell_data[cell_type] = {
-                string.ascii_lowercase[k]: numpy.random.rand(num_cells, dim)
+                string.ascii_lowercase[k]:
+                numpy.random.rand(num_cells, dim).astype(dtype)
                 for k in range(num_tags)
                 }
